pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "entgenai"
version = "1.0.0"
description = "Enterprise Generative AI Framework"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
metrics = ["grafana-api", "appdynamics"]